requirements.txt
openai-whisper
flask
boto3
flask-socketio
pyaudio
eventlet
awscli
python-dotenv
faster-whisper
webrtcvad
gc-python-utils